add readme (and totally nothing else)
